- run tpcds - add chowMin exact MIS

+++ src/main/java/omni/synopses/chowMin/TWODCM.java
@@ -156,8 +156,8 @@ public class TWODCM {
 //        return mix(attrValue, seed, width);
 //    }
 
-    public int memory() {
-        return depth  * widthS * widthT * 32;
+    public long memory() {
+        return depth  * widthS * widthT * 32L;
     }
